play-extra-parameter.html

<!doctype html>
<meta charset="utf-8">
<title>HTMLMediaElement: play() with extra paramenter</title>
<script src="support/testharness.js"></script>
<script src="support/testharnessreport.js"></script>
<script src="support/media.js"></script>
<div id="log"></div>
<script>
async_test(function (t) {
  var v = document.createElement("video");
  document.body.appendChild(v);
  v.src = getVideoURI("support/movie_5");
  try {
    v.play("one-more-param");
  } catch (e) {
    if (e && e.name && e.message)
      assert_unreached("Exception (" + e.name + ": " + e.message + ")");
    else if (e && e.message)
      assert_unreached("Exception (" + e.message + ")");
    else
      assert_unreached("Exception");
    t.done();
  }
  v.addEventListener("timeupdate", t.step_func(function () {
    assert_false(v.paused);
    if (v.currentTime > 0) {
      t.done();
    }
  }));
});
</script>
